interface: Fix argument count handling in inline cache
Before this change, when caching a rule invocation expression, the
inline cache did not take into account the number of arguments passed
to the invoked rule. As a result an invocation of a zero parameter
rule `foo'

  (foo)

followed by an invocation of the same rule with a different number of
arguments

  (foo 'bar)

would result in a single entry in the inline cache. This entry would
behave incorrectly for the second invocation by ignoring the argument
and invoking the rule without any arguments (instead of signaling an
error since rules take a fixed number of arguments).